MAINT: Move transerferfunction into color checker - there was no reason for the split at the moment.
