+++ src/daria/corrections/color/__init__.py
@@ -6,4 +6,3 @@ isort:skip_file
 
 
 from daria.corrections.color.colorchecker import *
-from daria.corrections.color.transferfunctions import *
